// src/app/models/masters-mpc/multimediause.ts
export class multimediause {
  id: number;
  name: string;
  description: string;
  maxAmount: number;
  color: string;
  createdByUserId: number;
  createdByUser: string;
  updatedByUserId: number;
  updatedByUser: string;
  createdDate: Date;
  updatedDate: Date;
  active: boolean;
  initialSetup: boolean;
}